Move mock handlers to mock and implement proxy. Signed-off-by: Will Piers <wpiers0405@gmail.com>
